# docs/openapi/frontend-surface.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FERN Frontend Surface",
    "version": "2026-03-31",
    "description": "Source-derived OpenAPI contract for the current frontend-facing gateway surface. It includes endpoints actively used by the frontend app or intentionally exposed for frontend diagnostics. It excludes internal control-plane routes, but now includes the auth session lifecycle, shared paged staff list envelopes with total-count metadata, the stock-safe sales, payroll, and procurement actions already exercised by the frontend, the dedicated cashier POS workspace at exact /pos built on existing protected sales and product endpoints, the protected non-cashier staff order queue at exact /order, and the public customer ordering plus public order-status flow under /api/v1/sales/public."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8180",
      "description": "Local gateway"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/org/outlets": {
      "get": {
        "summary": "List outlets",
        "parameters": [
          {
            "name": "regionId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Outlet list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OutletView"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/org/hierarchy": {
      "get": {
        "summary": "Region and outlet hierarchy",
        "responses": {
          "200": {
            "description": "Region and outlet hierarchy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgHierarchyView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/inventory/stock-balances": {
      "get": {
        "summary": "List stock balances",
        "parameters": [
          {
            "name": "outletId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "lowOnly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stock balance list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockBalanceView"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/inventory/transactions": {
      "get": {
        "summary": "List inventory transactions",
        "parameters": [
          {
            "name": "outletId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "itemId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inventory transaction list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryTransactionView"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/v1/procurement/goods-receipts": {
      "get": {
        "summary": "List goods receipts",
        "parameters": [
          {
            "name": "outletId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "poId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Goods-receipt list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoodsReceiptListPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/procurement/goods-receipts/{receiptId}": {
      "get": {
        "summary": "Get goods-receipt detail",
        "parameters": [
          {
            "name": "receiptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Goods-receipt detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoodsReceiptView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/procurement/goods-receipts/{receiptId}/approve": {
      "post": {
        "summary": "Approve a goods receipt",
        "parameters": [
          {
            "name": "receiptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Approved goods-receipt detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoodsReceiptView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/procurement/goods-receipts/{receiptId}/post": {
      "post": {
        "summary": "Post a goods receipt",
        "parameters": [
          {
            "name": "receiptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Posted goods-receipt event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoodsReceiptPostedEvent"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid bearer token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorPayload"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not allowed to access the requested scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorPayload"
            }
          }
        }
      },
      "NotFound": {
        "description": "Requested resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorPayload"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorPayload": {
        "type": "object",
        "required": [
          "timestamp",
          "error",
          "message"
        ],
        "properties": {
          "timestamp": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {}
        }
      },
...
      "RegionView": {
        "type": "object",
        "required": [
          "id",
          "code",
          "parentRegionId",
          "currencyCode",
          "name",
          "timezoneName"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "parentRegionId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "currencyCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "taxCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "timezoneName": {
            "type": "string"
          }
        }
      },
      "OutletView": {
        "type": "object",
        "required": [
          "id",
          "regionId",
          "code",
          "name",
          "status"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "regionId": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "address": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "openedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "closedAt": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "OrgHierarchyView": {
        "type": "object",
        "required": [
          "regions",
          "outlets"
        ],
        "properties": {
          "regions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegionView"
            }
          },
          "outlets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OutletView"
            }
          }
        }
      },
...
      "StockBalanceView": {
        "type": "object",
        "required": [
          "outletId",
          "itemId",
          "qtyOnHand",
          "unitCost",
          "updatedAt"
        ],
        "properties": {
          "outletId": {
            "type": "integer",
            "format": "int64"
          },
          "itemId": {
            "type": "integer",
            "format": "int64"
          },
          "qtyOnHand": {
            "type": "number"
          },
          "unitCost": {
            "type": "number"
          },
          "lastCountDate": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "InventoryTransactionView": {
        "type": "object",
        "required": [
          "id",
          "outletId",
          "itemId",
          "qtyChange",
          "businessDate",
          "txnTime",
          "txnType",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "outletId": {
            "type": "integer",
            "format": "int64"
          },
          "itemId": {
            "type": "integer",
            "format": "int64"
          },
          "qtyChange": {
            "type": "number"
          },
          "businessDate": {
            "type": "string"
          },
          "txnTime": {
            "type": "string"
          },
          "txnType": {
            "type": "string"
          },
          "unitCost": {
            "type": [
              "number",
              "null"
            ]
          },
          "createdByUserId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          }
        }
      },
...
      "GoodsReceiptItemView": {
        "type": "object",
        "required": [
          "id",
          "itemId",
          "uomCode",
          "qtyReceived",
          "unitCost",
          "lineTotal"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "itemId": {
            "type": "integer",
            "format": "int64"
          },
          "uomCode": {
            "type": "string"
          },
          "qtyReceived": {
            "type": "number"
          },
          "unitCost": {
            "type": "number"
          },
          "lineTotal": {
            "type": "number"
          },
          "manufactureDate": {
            "type": [
              "string",
              "null"
            ]
          },
          "expiryDate": {
            "type": [
              "string",
              "null"
            ]
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "GoodsReceiptView": {
        "type": "object",
        "required": [
          "id",
          "poId",
          "currencyCode",
          "receiptTime",
          "businessDate",
          "status",
          "totalPrice",
          "createdAt",
          "updatedAt",
          "items"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "poId": {
            "type": "integer",
            "format": "int64"
          },
          "currencyCode": {
            "type": "string"
          },
          "receiptTime": {
            "type": "string"
          },
          "businessDate": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "totalPrice": {
            "type": "number"
          },
          "supplierLotNumber": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdByUserId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "approvedByUserId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "approvedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoodsReceiptItemView"
            }
          }
        }
      },
      "GoodsReceiptListItemView": {
        "type": "object",
        "required": [
          "id",
          "poId",
          "outletId",
          "currencyCode",
          "receiptTime",
          "businessDate",
          "status",
          "totalPrice",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "poId": {
            "type": "integer",
            "format": "int64"
          },
          "outletId": {
            "type": "integer",
            "format": "int64"
          },
          "currencyCode": {
            "type": "string"
          },
          "receiptTime": {
            "type": "string"
          },
          "businessDate": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "totalPrice": {
            "type": "number"
          },
          "supplierLotNumber": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdByUserId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "approvedByUserId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "approvedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "GoodsReceiptPostedEvent": {
        "type": "object",
        "required": [
          "goodsReceiptId",
          "purchaseOrderId",
          "supplierId",
          "outletId",
          "businessDate",
          "currencyCode",
          "lineItems",
          "totalPrice",
          "postedAt"
        ],
        "properties": {
          "goodsReceiptId": {
            "type": "integer",
            "format": "int64"
          },
          "purchaseOrderId": {
            "type": "integer",
            "format": "int64"
          },
          "supplierId": {
            "type": "integer",
            "format": "int64"
          },
          "outletId": {
            "type": "integer",
            "format": "int64"
          },
          "businessDate": {
            "type": "string",
            "format": "date"
          },
          "currencyCode": {
            "type": "string"
          },
          "lineItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoodsReceiptItemView"
            }
          },
          "totalPrice": {
            "type": "number"
          },
          "postedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PagedResultMeta": {
        "type": "object",
        "required": [
          "limit",
          "offset",
          "totalCount",
          "hasNextPage"
        ],
        "properties": {
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "totalCount": {
            "type": "integer",
            "format": "int64"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        }
      },
...
      "GoodsReceiptListPage": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PagedResultMeta"
          },
          {
            "type": "object",
            "required": [
              "items"
            ],
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/GoodsReceiptListItemView"
                }
              }
            }
          }
        ]
      },
...
    }
  }
}